Add files via upload
This is a Number Guessing Game in which the system generates the random number within the range of 1 to 100 inclusively and points are allotted according to the guessed number and the generated number. The maximum point is 10 that is allotted when the guessed number and the system generated number is same and the minimum point is 0 allotted when the difference of guessed number and generated number is more than 40. 1,2,3,4,5 points are allotted according to how much the guessed number is near to generated number.