Employee Payroll System
This project is used to store all the necessary data of an employee such as employee id, attendance and calculating the salary according to the number of days employee attended to the company. In the project we can add, delete an employee and can update details of employee if any mistakes , update salary of an employee when employee gets increment in the salary.